Reset deployed repo to origin's HEAD. Closes #3
This fixes a problem with fast checkouts deploying origin/master (the
default). The old behavior reset to the deployed repo's HEAD, which is not
updated until `git branch`. Thus the old code was exported. Deploying a
second time would export the intended code.

The fix is to reset the deployed repo to origin's HEAD.

Deploying branches or tags continues to work as intended. E.g.

 set :revision, "origin/my_branch"
 set :revision "my_tag"